Refactoring Menu components #1
Make the menu components more flexible and
customizable. MenuItem component now is a
preact-router Link component implementation.
Handle the active MenuItem adding a rule
to add the active class to it.
