+++ src/menu/components/index.js
@@ -1,2 +1,3 @@
-export { default as Close } from './close'
-export { default as Item } from './item'
+export { default as MenuClose } from './menu-close'
+export { default as MenuItem } from './menu-item'
+export { default as Menu } from './menu.connected'
