--- src/menu/components/item.js
@@ -1,7 +1,0 @@
-import { h } from 'preact' /** @jsx h */
-
-export default ({ children }) => (
-  <div className="components--menu-item">
-    {children}
-  </div>
-)
