+++ src/menu/components/menu-item.js
@@ -1,0 +1,16 @@
+import { h } from 'preact' /** @jsx h */
+import { Link } from 'preact-router'
+
+export default ({ children, href }) => {
+  const dynamicClassNames = []
+  if (window.location.pathname === href) dynamicClassNames.push('active')
+
+  return (
+    <Link
+      className={`components--menu-item ${dynamicClassNames.join(' ')}`}
+      href={href}
+    >
+      {children}
+    </Link>
+  )
+}
