+++ src/menu/components/menu.js
@@ -1,14 +1,12 @@
 import { h } from 'preact' /** @jsx h */
+import MenuClose from './menu-close'
 
-import Close from './close'
-import Item from './item'
-
-export default ({ active, setActive }) => !active ? null : (
-  <div className="components--menu">
-    <Close onClick={() => setActive(false)} />
-    <Item>Início</Item>
-    <Item>Criar Homenagem</Item>
-    <Item>Editar Homenagem</Item>
-    <Item>Quem Somos</Item>
-  </div>
-)
+export default ({ children, active, setActive, ...props }) => {
+  console.log('props', props)
+  return !active ? null : (
+    <nav className="components--menu">
+      <MenuClose onClick={() => setActive(false)} />
+      {children}
+    </nav>
+  )
+}
