+++ src/styles/all.sass
@@ -13,9 +13,9 @@
 @import 'base/doc'
 
 // menu module
-@import 'menu/index'
-@import 'menu/close'
-@import 'menu/item'
+@import 'menu/menu'
+@import 'menu/menu-close'
+@import 'menu/menu-item'
 
 // tags first; pages can customize
 @import 'tags/header'
