+++ src/styles/menu/menu-item.sass
@@ -1,8 +1,14 @@
 @import url('https://fonts.googleapis.com/css?family=Roboto:300')
 
 .components--menu-item
+  display: block
   font-family: Roboto
   font-weight: 300
   font-size: 24px
   color: #FFFFFF
   line-height: 55px
+  padding: 0 16px
+  text-decoration: none
+
+  &.active
+    background-color: rgba(255,255,255,.1)
