+++ src/styles/menu/menu.sass
@@ -5,6 +5,5 @@
   bottom: 0
   left: 0
   width: 300px
-  padding: 16px
-  padding-top: 90px
+  padding-top: 106px
   z-index: 5
